+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # unreleased
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
+* Fix issue where environment was not automatically converted from a string to a symbol (#148, #164)
 
 # 2.90.0
 * Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.